+++ ntfy.xcodeproj/xcuserdata/pheckel.xcuserdatad/xcschemes/xcschememanagement.plist
@@ -28,12 +28,12 @@
 		<key>ntfy.xcscheme_^#shared#^_</key>
 		<dict>
 			<key>orderHint</key>
-			<integer>1</integer>
+			<integer>0</integer>
 		</dict>
 		<key>ntfyNSE.xcscheme_^#shared#^_</key>
 		<dict>
 			<key>orderHint</key>
-			<integer>0</integer>
+			<integer>1</integer>
 		</dict>
 	</dict>
 </dict>
